requirements.txt
boto3==1.34.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.26.0
flask==3.0.0
matplotlib==3.8.0
seaborn==0.13.0
joblib==1.3.2
requests==2.31.0
streamlit==1.37.0
python-dotenv